- Bump update for win32 build.
git-svn-id: https://svn.linuxsampler.org/svn/qsampler/trunk@1671 d2c4447f-3961-4fff-bec3-96d84be9abe5

+++ src/qsamplerChannelFxForm.cpp
@@ -258,7 +258,7 @@ void ChannelFxForm::onFxSendSelection(const QModelIndex& index) {
 
 	m_ui.depthCtrlComboBox->setCurrentIndex(pFxSend->sendDepthMidiCtrl());
 	m_ui.depthSpinBox->setValue(
-		::round(pFxSend->currentDepth() * 100.0)
+		int(::round(pFxSend->currentDepth() * 100.0))
 	);
 }
